Implement add new doc putfile

+++ app/src/main/java/com/graphitedocs/graphitedocs/utils/models/DocsListItem.kt
@@ -1,5 +1,5 @@
 package com.graphitedocs.graphitedocs.utils.models
 
 class DocsListItem(var title: String, var date: String, var sharedWith: List<String>?,
-                   var tags: List<String>?, var author: String, var id: Int,
+                   var tags: List<String>?, var author: String, var id: Long,
                    var updated: String?, var created: String?)
